+++ src/ska_tmc_dishleafnode/manager/health_rules.py
@@ -9,7 +9,7 @@ HEALTH_RULES = {
     HealthState.OK: [
         Rule("kvalue_validation_result == 'OK'"),
     ],
-    HealthState.DEGRADED: [
+    HealthState.FAILED: [
         Rule("kvalue_validation_result != 'OK'"),
     ],
 }