+++ CHANGELOG.md
@@ -5,6 +5,14 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+[0.27.3]
+********
+Fixed
+-----
+* Added event subscription for Dish kValue updates
+* Implemented validation rules using the rule engine
+* Updated dish healthState based on kValue validation results
+
 [0.27.2]
 ********
 Fixed